refactor/gui-v2-store-form-layout-improved-and-cleanup
